==> Scripts/1. Cleaning Orders, Products, Department Data Sets.ipynb <==
#import libraries
import pandas as pd
import numpy as np
import os

path = '../Data/Raw Data'

df_ord = pd.read_csv(path + '/orders.csv')

df_ord.shape

df_ord.head()

# drop uneccessary column eval_set
df_ord = df_ord.drop(columns = ['eval_set'])

df_ord.head()

# changing the data type for columns with unique values and numbers that equal a day of the week
df_ord[['order_id', 'user_id', 'order_number', 'order_dow']] = df_ord[['order_id', 'user_id', 'order_number', 'order_dow']].astype('str')

df_ord.info()

# checking for missing data in Orders dataset
missing_values = df_ord.isnull().sum()
print(missing_values)

# using the average number of days since prior order to impute that for the 206209 missing values in that column
average_value = df_ord['days_since_prior_order'].mean()
print(average_value)

#Imputing the value of 11 in the missing 206209 values
impute_value = 11
df_ord['days_since_prior_order'].fillna(impute_value, inplace = True)

missing_values = df_ord.isnull().sum()
print(missing_values)

# Finding duplicates in orders data set
duplicate_rows = df_ord[df_ord.duplicated()]
print(duplicate_rows)

# saving clean orders dataset
path = '../Data/Clean Data'
df_ord.to_csv(os.path.join(path, 'orders_clean'))

# Cleaning Products Data Set

del df_ord

path = '../Data/Raw Data'

df_prod = pd.read_csv(path + '/products.csv')

df_prod.shape

df_prod.head()

# dropping unnecessary column aisle_id as this project is for online shoppping
df_prod = df_prod.drop(columns = ['aisle_id'])

df_prod.head()

# finding missing values 
missing_values = df_prod.isnull().sum()
print(missing_values)

# Deleting the 16 missing values
df_nan = df_prod[df_prod['product_name'].isnull() == True]

df_prod2 = df_prod[df_prod['product_name'].isnull() == False]

df_nan = df_prod2

df_prod2 = df_prod[df_prod['product_name'].isnull() == False]

missing_values = df_prod2.isnull().sum()
print(missing_values)

df_prod.info()

del df_prod

# Finding duplicates in products data set
df_dups = df_prod2[df_prod2.duplicated()]

df_dups

# deleting duplicates
df_prod3 = df_prod2.drop_duplicates()

df_prod3.shape

del df_prod2

# Looking for outliers on prices
max_value = df_prod3['prices'].nlargest(10)
print(max_value)

# looking into rows with $9999 and $14900 prices
max_price_row = df_prod3.nlargest(5, 'prices')
print(max_price_row)

# inputting the average US cost of 2% reduced fat milk: 4
row_index = 33666
impute_value = 4
df_prod3.at[row_index, 'prices'] = impute_value

# imputting the average US cost of lowfat 2% milkfat cottage cheese
row_index = 21554
impute_value = 4
df_prod3.at[row_index, 'prices'] = impute_value

max_price_row = df_prod3.nlargest(5, 'prices')
print(max_price_row)

# saving clean products dataset
path = '../Data/Clean Data'
df_prod3.to_csv(os.path.join(path, 'products_clean'))

del df_prod3

# Cleaning Departments Data set

path = '../Data/Raw Data'

df_dep = pd.read_csv(path + '/departments.csv')

df_dep.head()

df_dep = df_dep.T

df_dep.head()

# channging the first row to the header of columns
df_dep.columns = df_dep.iloc[0]

df_dep = df_dep.drop(df_dep.index[0])

df_dep

path = '../Data'

df_dep.to_csv(os.path.join(path, 'Clean Data', 'department_clean.csv'))

del df_dep
