==> Scripts/3. Merging Orders and Order Products Prior.ipynb <==
import pandas as pd
import numpy as np
import os
from os.path import join

path = '../Data'

df_opr = pd.read_csv(join(path, 'Clean Data', 'clean_order_products_prior.csv'))

df_opr.head()

# dropping Unnammed:0 column
df_opr = df_opr.drop(columns = ['Unnamed: 0'])

df_opr.head()

df_ord = pd.read_csv(join(path, 'Clean Data', 'orders_clean'))

df_ord.head()

# dropping Unnammed: 0 column
df_ord = df_ord.drop(columns = ['Unnamed: 0'])

df_ord.head()

df_opr.shape

df_ord.shape

# merging orders and order products prior together
df_merg = df_ord.merge(df_opr, on = 'order_id', how = 'inner', indicator = True)

df_merg.head()

# checking merge results
value_counts = df_merg['_merge'].value_counts()
print(value_counts)

# checking for missing values
missing_values = df_merg.isnull().sum()
print(missing_values)

# deleting merge column
df_merg = df_merg.drop(columns = ['_merge'])

del df_ord

del df_opr

# Checking for duplicate values
duplicate_rows = df_merg[df_merg.duplicated()]
if duplicate_rows.shape[0]==0:
    print('No Duplicates Found')
else:
    print(duplicate_rows.head())

df_merg.head()

# saving merged dataset
import pickle
df_merg.to_pickle(join(path, 'Clean Data', 'orders_products_prior_merge.pkl'))
