==> Scripts/4. Merging Orders, Products, Order Products Prior with Products and Department Data Sets.ipynb <==
import pandas as pd
import numpy as np
import os
from os.path import join
import pickle

path = '../Data'

df_pro = pd.read_csv(join(path, 'Clean Data', 'products_clean'))

df_main = pd.read_pickle(join(path, 'Clean Data', 'orders_products_prior_merge.pkl'))

df_pro.head()

# drop Unnamed: 0 column
df_pro = df_pro.drop(columns = ['Unnamed: 0'])

df_pro.head()

df_main.head()

df_pro.shape

df_main.shape

# merging the main data set with products
df_main2 = df_main.merge(df_pro, on = 'product_id', how = 'inner', indicator = True)

df_main2.head()

# checking merge results
value_counts = df_main2['_merge'].value_counts()
print(value_counts)

# Checking for duplicate values
duplicate_rows = df_main2[df_main2.duplicated()]
if duplicate_rows.shape[0]==0:
    print('No Duplicates Found')
else:
    print(duplicate_rows.head())

# checking for missing values
missing_values = df_main2.isnull().sum()
print(missing_values)

#dropping merge column
df_main2 = df_main2.drop(columns = ['_merge'])

df_main2.head()

del df_pro

del df_main

# saving main2 data set
df_main2.to_pickle(join(path, 'Clean Data', 'orders_prior_products_combined.pkl'))

# Next Data set to merge: department
df_dep = pd.read_csv(join(path, 'Clean Data', 'department_clean.csv'))

df_dep.head()

df_dep = df_dep.rename(columns = {'Unnamed: 0': 'department_id'})

df_dep.head()

# merging departments with 2nd main dataset
df_main3 = df_main2.merge(df_dep, on = 'department_id', how = 'inner', indicator = True)

df_main3.head()

# checking main3 results
value_counts = df_main3['_merge'].value_counts()
print(value_counts)

# Checking for duplicate values
duplicate_rows = df_main3[df_main3.duplicated()]
if duplicate_rows.shape[0]==0:
    print('No Duplicates Found')
else:
    print(duplicate_rows.head())

# checking for missing values
missing_values = df_main3.isnull().sum()
print(missing_values)

# dropping the merge column
df_main3 = df_main3.drop(columns = ['_merge'])

df_main3.head()

# saving main3 data set
df_main3.to_pickle(join(path, 'Clean Data', 'orders_prior_prods_dep.pkl'))

del df_main2

del df_main3

import gc
gc.collect()
