==> Scripts/6. Downsizing the Final Data Set.ipynb <==
import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy
import pickle
from os.path import join

path = '../Data'

df = pd.read_pickle(join(path, 'Clean Data', 'main_dataset.pkl'))

df.head()

df.shape

# Client wishes to exclude customer data for users that have less than 5 orders.
df['order_act'] = df.groupby('user_id')['order_number'].transform(lambda x: 'Low Activity' if x.max() <= 5 else 'High Activity')

df.head()

df['order_act'].value_counts()

df1 = df[df['order_act']!= 'Low Activity']

df1.shape

del df

# creating sample data set to use moving forward.
np.random.seed(4)
dev = np.random.rand(len(df1)) <= 0.7

# Store 70% and 30% of the sample in dataframe big and small
big = df1[dev]
small = df1[~dev]

df2 = big

df2.shape

df3 = small

df3.shape

#saving data sets big and small
df2.to_pickle(join(path, 'Clean Data', 'final_data_big.pkl'))

df3.to_pickle(join(path, 'Clean Data', 'final_data_small.pkl'))

del df2

del df3
