==> Scripts/7. Column Derivations.ipynb <==
import pandas as pd
import numpy as np
import os
import pickle
from os.path import join

path = '../Data'

# Moving forward using the 70% sample of the final data set
df = pd.read_pickle(join(path, 'Clean Data', 'final_data_big.pkl'))

df.shape

df.head()

# Looking into the order day of week
df['order_dow'].value_counts(dropna = False)

# creating a new column for busiest days of week
def profiles(row):
    if row['order_dow'] == 0:
        return 'Busiest Day'
    elif row['order_dow'] == 4:
        return 'Least Busy Day'
    elif row['order_dow'] == 1:
        return '2nd Busiest Day'
    elif row['order_dow'] ==3:
        return '2nd Least Busy Day'
    else:
        return 'Regularly Busy Day'
df['busiest_days'] = df.apply(profiles, axis = 1)

df.head()

freq = df['order_hour_of_day'].value_counts()
print(freq)

# adding column from order_hour_of_day to show when the most, average, fewest orders appear.
def profiles(row):
    if row['order_hour_of_day'] in [10,11,14,15,13,12,16,9]:
        return 'Most Orders'
    elif row['order_hour_of_day'] in [17,8,18,19,20,7,21,22]:
        return 'Average Orders'
    else:
        return 'Fewest Orders'

df['busiest_period_of_day'] = df.apply(profiles, axis = 1)

df.head()

df['busiest_period_of_day'].value_counts(dropna = False)

# Trying some groupby coding
df.groupby('product_name')

df.groupby('department_id').agg({'order_number' : ['mean']})

df.groupby('department_id').agg({'order_number': ['mean', 'min', 'max']})

# adding a column for max number of orders
df['max_order'] = df.groupby(['user_id'])['order_number'].transform(np.max)

df.head()

# adding column for Customer Loyalty
df.loc[df['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

df.loc[(df['max_order'] <= 40) & (df['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

df.loc[df['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

df['loyalty_flag'].value_counts(dropna = False)

# showing statistics for loyalty flag
df.groupby('loyalty_flag').agg({'prices': ['mean', 'min', 'max']})

#creating a column for average prices by user id
df['avg_prices'] = df.groupby(['user_id'])['prices'].transform(np.mean)

df.head()

# creating flags in new column spender
df.loc[df['avg_prices'] < 10, 'spender'] = 'Low spender'

df.loc[df['avg_prices'] >= 10, 'spender'] = 'High spender'

df.head()

df['spender'].value_counts(dropna = False)

# creating new column for average frequency of order
df['avg_freq'] = df.groupby(['user_id'])['days_since_prior_order'].transform(np.mean)

df.head()

# creating order frequency flags from average frequency column
df.loc[df['avg_freq'] > 20, 'order_freq'] = 'Non-frequent Customer'

df.loc[(df['avg_freq'] >10) & (df['avg_freq'] <=20), 'order_freq'] = 'Reguler Customer'

df.loc[df['avg_freq'] <=10, 'order_freq'] = 'Frequent Customer'

df['order_freq'].value_counts(dropna = False)

df.to_pickle(join(path, 'Clean Data', 'final_dataset.pkl'))

del df
