==> Scripts/8. Running Stats.ipynb <==
import pandas as pd
import numpy as np
import os
import pickle
from os.path import join

path = '../Data'

df = pd.read_pickle(join(path, 'Clean Data', 'final_dataset.pkl'))

df.head()

# creating a column that shows the total cost per order per customer
total_costs = df.groupby(['user_id', 'order_id'])['prices'].sum().reset_index(name='cost_per_order')

df = df.merge(total_costs, on = ['user_id', 'order_id'])

df.head()

# Stats for Income
average = df['income'].mean()
mode = df['income'].mode().values[0]
minimum = df['income'].min()
maximum = df['income'].max()
print("Average:", average)
print("Mode:", mode)
print("Minimum:", minimum)
print("Maximum:", maximum)

# Stats for Number of Dependents
average = df['n_dependants'].mean()
mode = df['n_dependants'].mode().values[0]
minimum = df['n_dependants'].min()
maximum = df['n_dependants'].max()
print("Average:", average)
print("Mode:", mode)
print("Minimum:", minimum)
print("Maximum:", maximum)

# Value counts for Profile
df['profile'].value_counts()

# Stats for Average Frequency of order (in days)
average = df['avg_freq'].mean()
mode = df['avg_freq'].mode().values[0]
minimum = df['avg_freq'].min()
maximum = df['avg_freq'].max()
print("Average:", average)
print("Mode:", mode)
print("Minimum:", minimum)
print("Maximum:", maximum)

# Stats for Age
average = df['Age'].mean()
mode = df['Age'].mode().values[0]
minimum = df['Age'].min()
maximum = df['Age'].max()
print("Average:", average)
print("Mode:", mode)
print("Minimum:", minimum)
print("Maximum:", maximum)

# value counts for region
df['region'].value_counts()

# value counts for spender
df['spender'].value_counts()

# Stats for Order hour of day
average = df['order_hour_of_day'].mean()
mode = df['order_hour_of_day'].mode().values[0]
minimum = df['order_hour_of_day'].min()
maximum = df['order_hour_of_day'].max()
print("Average:", average)
print("Mode:", mode)
print("Minimum:", minimum)
print("Maximum:", maximum)

# Stats for the 10 most purchased items 
top_10 = df['product_name'].value_counts().head(10)
print(top_10)

# Stats for the 10 least purchased items
bottom_10 = df['product_name'].value_counts().tail(10)
print(bottom_10)

# value counts for order frequency
df['order_freq'].value_counts()

# saving data set
df.to_pickle(join(path, 'Clean Data'))
